% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice.R
\name{activeRmd_KnitAddJsServe}
\alias{activeRmd_KnitAddJsServe}
\title{Knit active Rmd, add JS and serve based on params$after_body}
\usage{
activeRmd_KnitAddJsServe()
}
\value{

}
\description{
Knit active Rmd, add JS and serve based on params$after_body
}
\examples{
none
}
