% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice.R
\name{copyServeChunk}
\alias{copyServeChunk}
\title{Have web serve R chunk in your clipboard for you to paste}
\usage{
copyServeChunk()
}
\value{

}
\description{
Have web serve R chunk in your clipboard for you to paste
}
\examples{
none
}
