% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice.R
\name{create_jqueryPage}
\alias{create_jqueryPage}
\title{Create a jquery experiemtal index page with body content from the copy of
 a webpage inspection element}
\usage{
create_jqueryPage()
}
\value{

}
\description{
Create a jquery experiemtal index page with body content from the copy of
 a webpage inspection element
}
\examples{
none
}
