% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webTemplate.R
\name{create_materializedFramework}
\alias{create_materializedFramework}
\title{Create a Rmd to support Materialized  frame work}
\usage{
create_materializedFramework()
}
\value{
No return
}
\description{
Create a Rmd to support Materialized  frame work
}
\examples{
none
}
