% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{extract_withinbodyHtml}
\alias{extract_withinbodyHtml}
\title{Extract withinbody html components from myChoice}
\usage{
extract_withinbodyHtml(myChoice)
}
\arguments{
\item{myChoice}{A list from menu<- get_menu(); myChoice <- list(menu$...(), menu$...(),...)}
}
\value{
A taglist
}
\description{
Extract withinbody html components from myChoice
}
\examples{
none
}
