% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{generate_carousel_indicators}
\alias{generate_carousel_indicators}
\title{Generate carousel with indicators}
\usage{
generate_carousel_indicators(id, listOfImgUrls)
}
\arguments{
\item{id}{A character}

\item{listOfImgUrls}{A character vector of urls}
}
\value{

}
\description{
Generate carousel with indicators
}
\examples{
id="carouselExampleIndicators"
listOfImgUrls <- list(
  "https://i.imgur.com/XPcc1WF.png",
  "https://i.imgur.com/CY2AWmB.png"
)
generate_carousel_indicators(id, listOfImgUrls)
}
