% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{generate_template}
\alias{generate_template}
\title{Generate htmlcontext to be save_html}
\usage{
generate_template(myChoice, .body = NULL)
}
\arguments{
\item{myChoice}{A list. Use menu <- get_menu() to obtain menu, then construct myChoice= list( menu$optionM(), menu$optionK(),... )}

\item{.body}{A tagList of each body child elements other than afterbodyJs and afterbodyHtml. However, some framework such as google login requires a holder withinbody. You need to place it in your .body}
}
\value{
A shiny.tag class htmlContext ready to save_html
}
\description{
Generate htmlcontext to be save_html
}
\examples{
none
}
