% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{get_depTemplate}
\alias{get_depTemplate}
\title{Get an empty template to fill up for html dependency}
\usage{
get_depTemplate()
}
\value{

}
\description{
Get an empty template to fill up for html dependency
}
\examples{
none
}
