% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleLoginModule.R
\name{googleLoginModule}
\alias{googleLoginModule}
\title{Generate a google login environment, with preprocessing function to tailor developer app}
\usage{
googleLoginModule()
}
\value{

}
\description{
Generate a google login environment, with preprocessing function to tailor developer app
}
\examples{
none
}
