% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize.R
\name{materializeService}
\alias{materializeService}
\title{Service that generate elements under materialize framework}
\usage{
materializeService()
}
\value{

}
\description{
Service that generate elements under materialize framework
}
\examples{
none.
}
