% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice.R
\name{serveTheLatestHtml_ActiveFileFolder}
\alias{serveTheLatestHtml_ActiveFileFolder}
\title{serve the last modified html (usually the latest knitted Rmd) of the current Active File folder}
\usage{
serveTheLatestHtml_ActiveFileFolder(useLocalhost = T)
}
\arguments{
\item{useLocalhost}{A logical, default=T}
}
\value{

}
\description{
serve the last modified html (usually the latest knitted Rmd) of the current Active File folder
}
\examples{
none
}
