% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webTemplate.R
\name{create_materializedFramework}
\alias{create_materializedFramework}
\title{Create a R Markdown with file system that supports Materialized Web framework}
\usage{
create_materializedFramework()
}
\value{
No return. Simply a file system
}
\description{
Create a R Markdown with file system that supports Materialized Web framework
}
\examples{
none.
}
